revert veteran-address and move changes to new file
