+++ src/applications/disability-benefits/686c-674/config/address-schema.js
@@ -71,51 +71,6 @@ const MilitaryBaseInfo = () => (
   </div>
 );
 
-export const addressSchema = {
-  type: 'object',
-  properties: {
-    isMilitary: {
-      type: 'boolean',
-    },
-    'view:livesOnMilitaryBaseInfo': {
-      type: 'object',
-      properties: {},
-    },
-    country: {
-      type: 'string',
-      enum: countries.map(country => country.value),
-      enumNames: countries.map(country => country.label),
-    },
-    addressLine1: {
-      type: 'string',
-      minLength: 1,
-      maxLength: 100,
-      pattern: '^.*\\S.*',
-    },
-    addressLine2: {
-      type: 'string',
-      minLength: 1,
-      maxLength: 100,
-      pattern: '^.*\\S.*',
-    },
-    addressLine3: {
-      type: 'string',
-      minLength: 1,
-      maxLength: 100,
-      pattern: '^.*\\S.*',
-    },
-    city: {
-      type: 'string',
-    },
-    state: {
-      type: 'string',
-    },
-    postalCode: {
-      type: 'string',
-    },
-  },
-};
-
 /**
  * Builds address schema based on isMilitaryAddress.
  * @param {boolean} isMilitaryBaseAddress represents whether or not the form page requires the address to support the option of military address.
@@ -404,169 +359,6 @@ export const addressUISchema = (
             const countryName = get(countryNamePath, formData);
             return countryName === USA.value || !countryName;
           },
-        },
-      },
-    };
-  })();
-};
-
-const MILITARY_BASE_PATHII = 'isMilitary';
-
-export const addressUiSchemaII = (path, callback) => {
-  // As mentioned above, there are certain fields that depend on the values of other fields when using updateSchema, replaceSchema, and hideIf.
-  // The two constants below are paths used to retrieve the values in those other fields.
-  const livesOnMilitaryBasePath = `${path}.${MILITARY_BASE_PATHII}`;
-  const insertArrayIndex = (key, index) => key.replace('[INDEX]', `[${index}]`);
-  const checkBoxTitleState = path.includes('veteran') ? 'I' : 'They';
-
-  return (function returnAddressUI() {
-    return {
-      isMilitary: {
-        'ui:title': `${checkBoxTitleState} live on a United States military base outside of the U.S.`,
-      },
-      'view:livesOnMilitaryBaseInfo': {
-        'ui:description': MilitaryBaseInfo,
-      },
-      country: {
-        'ui:required': callback,
-        'ui:title': 'Country',
-        'ui:options': {
-          updateSchema: (formData, schema, uiSchema, index) => {
-            let militaryBasePath = livesOnMilitaryBasePath;
-            let countryPath = path;
-            if (typeof index === 'number') {
-              militaryBasePath = insertArrayIndex(
-                livesOnMilitaryBasePath,
-                index,
-              );
-              countryPath = insertArrayIndex(path, index);
-            }
-            const countryUI = uiSchema;
-            const countryFormData = get(countryPath, formData);
-            const livesOnMilitaryBase = get(militaryBasePath, formData);
-            if (livesOnMilitaryBase) {
-              countryUI['ui:disabled'] = true;
-              countryFormData.country = USA.value;
-              return {
-                enum: [USA.value],
-                enumNames: [USA.label],
-                default: USA.value,
-              };
-            }
-            countryUI['ui:disabled'] = false;
-            return {
-              type: 'string',
-              enum: countries.map(country => country.value),
-              enumNames: countries.map(country => country.label),
-            };
-          },
-        },
-      },
-      addressLine1: {
-        'ui:required': callback,
-        'ui:title': 'Street',
-        'ui:errorMessages': {
-          required: 'Street address is required',
-          pattern: 'Street address must be under 100 characters',
-        },
-      },
-      addressLine2: {
-        'ui:title': 'Line 2',
-      },
-      addressLine3: {
-        'ui:title': 'Line 3',
-      },
-      city: {
-        'ui:required': callback,
-        'ui:errorMessages': {
-          required: 'City is required',
-          pattern: 'City must be under 100 characters',
-        },
-        'ui:options': {
-          replaceSchema: (formData, schema, uiSchema, index) => {
-            let militaryBasePath = livesOnMilitaryBasePath;
-            if (typeof index === 'number') {
-              militaryBasePath = insertArrayIndex(
-                livesOnMilitaryBasePath,
-                index,
-              );
-            }
-            const livesOnMilitaryBase = get(militaryBasePath, formData);
-            if (livesOnMilitaryBase) {
-              return {
-                type: 'string',
-                title: 'APO/FPO/DPO',
-                enum: militaryCities.map(city => city.value),
-                enumNames: militaryCities.map(city => city.label),
-              };
-            }
-            return {
-              type: 'string',
-              title: 'City',
-              minLength: 1,
-              maxLength: 100,
-              pattern: '^.*\\S.*',
-            };
-          },
-        },
-      },
-      state: {
-        'ui:required': (formData, index) => {
-          let countryPath = `${path}.country`;
-          if (typeof index === 'number') {
-            countryPath = insertArrayIndex(countryPath, index);
-          }
-          const livesOnMilitaryBase = get(livesOnMilitaryBasePath, formData);
-          const country = get(countryPath, formData);
-          return (country && country === USA.value) || livesOnMilitaryBase;
-        },
-        'ui:errorMessages': {
-          required: 'Please enter a valid State, Province, or Region',
-        },
-        'ui:options': {
-          replaceSchema: (formData, schema, uiSchema, index) => {
-            let militaryBasePath = livesOnMilitaryBasePath;
-            let countryPath = `${path}.country`;
-            if (typeof index === 'number') {
-              militaryBasePath = insertArrayIndex(
-                livesOnMilitaryBasePath,
-                index,
-              );
-              countryPath = insertArrayIndex(countryPath, index);
-            }
-            const livesOnMilitaryBase = get(militaryBasePath, formData);
-            const country = get(countryPath, formData);
-            if (livesOnMilitaryBase) {
-              return {
-                type: 'string',
-                title: 'State',
-                enum: Object.keys(MILITARY_STATES),
-                enumNames: Object.values(MILITARY_STATES),
-              };
-            } else if (!livesOnMilitaryBase && country === 'USA') {
-              return {
-                type: 'string',
-                title: 'State',
-                enum: states50AndDC.map(state => state.value),
-                enumNames: states50AndDC.map(state => state.label),
-              };
-            } else {
-              return {
-                type: 'string',
-                title: 'State/Province/Region',
-              };
-            }
-          },
-        },
-      },
-      postalCode: {
-        'ui:required': callback,
-        'ui:title': 'Postal Code',
-        'ui:errorMessages': {
-          required: 'Postal code is required',
-        },
-        'ui:options': {
-          widgetClassNames: 'usa-input-medium',
         },
       },
     };
