+++ src/applications/disability-benefits/686c-674/config/chapters/veteran-information/veteran-address/veteran-address.js
@@ -1,13 +1,15 @@
 import cloneDeep from 'platform/utilities/data/cloneDeep';
 import emailUI from 'platform/forms-system/src/js/definitions/email';
 import { veteranInformation } from '../../../utilities';
-import { addressSchema, addressUiSchemaII } from '../../../address-schema';
+import { buildAddressSchema, addressUISchema } from '../../../address-schema';
 
 const veteranContactInformationSchema = cloneDeep(
   veteranInformation.properties.veteranAddress,
 );
 
-veteranContactInformationSchema.properties.veteranAddress = addressSchema;
+veteranContactInformationSchema.properties.veteranAddress = buildAddressSchema(
+  true,
+);
 
 // add confirm email field on the frontend only
 veteranContactInformationSchema.properties['view:confirmEmail'] = {
@@ -23,7 +25,7 @@ export const schema = {
 
 export const uiSchema = {
   veteranContactInformation: {
-    veteranAddress: addressUiSchemaII(
+    veteranAddress: addressUISchema(
       'veteranContactInformation.veteranAddress',
       () => true,
     ),
